requirements.txt
networkx
# local package
-e .

# external requirements
tweepy
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
firebase-admin
google-cloud
python-modularity-maximization==0.0.1rc4
